[Тема: 14, Урок: 1/2]: Переименованы файлы / классы / протокол статистики в соответствии к ТЗ;
