--- MovieQuiz/Services/StatisticServiceProtocol.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-protocol StatisticServiceProtocol {
-    var gamesCount: Int { get }
-    var bestGame: GameResultModel { get }
-    var totalAccuracy: Double { get }
-    func store(correct count: Int, total amount: Int)
-}
